+++ lib/rails/resource_service.rb
@@ -1,4 +1,5 @@
 require 'facets/string/underscore'
+require 'glimmer/util/url_builder'
 
 module Rails
   # ResourceService supports index/show/create/update/destroy REST API calls for any Rails resources
@@ -8,19 +9,27 @@ module Rails
     TIMESTAMP_ATTRIBUTES = ['created_at', 'updated_at']
     
     class << self
-      def index(resource: nil, resource_class: nil, singular_resource_name: nil, plural_resource_name: nil, index_resource_url: nil, params: nil, &response_handler)
+      def index(resource: nil, resource_class: nil, singular_resource_name: nil, plural_resource_name: nil, index_resource_url: nil, params: nil, root: nil, &response_handler)
         resource_class ||= resource&.class
         singular_resource_name ||= singular_resource_name_for_resource_class(resource_class)
         plural_resource_name ||= "#{singular_resource_name}s"
         index_resource_url ||= "/#{plural_resource_name}.json"
-        HTTP.get(index_resource_url, payload: index_show_destroy_resource_params(params: params.to_h)) do |response|
+        root = plural_resource_name if root == true
+        index_resource_url = Glimmer::Util::UrlBuilder.new.url(index_resource_url).params(index_show_destroy_resource_params(params: params.to_h)).to_s
+        HTTP.get(index_resource_url) do |response|
           if response.ok? && !resource_class.nil?
             resource_response_objects = Native(response.body)
-            # TODO consider having index not make the assumpion that resource_response_objects is an array in case it isn't
-            # when including pagination information (maybe test if it is an array first)
-            resources = resource_response_objects.map { |resource_response_object| build_resource_from_response_object(resource_class:, resource_response_object:) }
+            metadata = {}
+            if root
+              resources = resource_response_objects[root].map { |resource_response_object| build_resource_from_response_object(resource_class:, resource_response_object:) }
+              populate_metadata_from_native_response(metadata:, native_response: resource_response_objects, root:)
+            else
+              resources = resource_response_objects.map { |resource_response_object| build_resource_from_response_object(resource_class:, resource_response_object:) }
+            end
           end
-          response_handler.call(response, resources)
+          puts 'metadata'
+          puts metadata
+          response_handler.call(response, resources, metadata) # TODO support 3rd argument for metadata
         end
       end
       
@@ -30,7 +39,8 @@ module Rails
         singular_resource_name ||= singular_resource_name_for_resource_class(resource_class)
         plural_resource_name ||= "#{singular_resource_name}s"
         show_resource_url ||= "/#{plural_resource_name}/#{resource_id}.json"
-        HTTP.get(show_resource_url, payload: index_show_destroy_resource_params(params: params.to_h)) do |response|
+        show_resource_url = Glimmer::Util::UrlBuilder.new.url(show_resource_url).params(index_show_destroy_resource_params(params: params.to_h)).to_s
+        HTTP.get(show_resource_url) do |response|
           if response.ok? && !resource_class.nil?
             resource_response_object = Native(response.body)
             resource = build_resource_from_response_object(resource_class:, resource_response_object:)
@@ -82,7 +92,8 @@ module Rails
         singular_resource_name ||= singular_resource_name_for_resource_class(resource_class)
         plural_resource_name ||= "#{singular_resource_name}s"
         destroy_resource_url ||= "/#{plural_resource_name}/#{resource_id}.json"
-        HTTP.delete(destroy_resource_url, payload: index_show_destroy_resource_params(params: params.to_h), &response_handler)
+        destroy_resource_url = Glimmer::Util::UrlBuilder.new.url(destroy_resource_url).params(index_show_destroy_resource_params(params: params.to_h)).to_s
+        HTTP.delete(destroy_resource_url, &response_handler)
       end
       
       def index_show_destroy_resource_params(params: nil)
@@ -112,6 +123,23 @@ module Rails
         resource = resource_class.new
         resource_response_object.each { |attribute, value| resource.send("#{attribute}=", value) rescue nil }
         resource
+      end
+      
+      def populate_metadata_from_native_response(metadata:, native_response:, root:)
+        native_response.each do |attribute, value|
+          next if attribute == root
+          begin
+            value.class
+            metadata[attribute] = value
+          rescue Exception => e
+            if `typeof #{value}` == 'object'
+              metadata[attribute] = {}
+              populate_metadata_from_native_response(metadata: metadata[attribute], native_response: Native(value), root: nil)
+            else
+              metadata[attribute] = value
+            end
+          end
+        end
       end
       
       # TODO consider offering generic Rails::ResourceService.get, Rails::ResourceService.put, Rails::ResourceService.post, Rails::ResourceService.patch, Rails::ResourceService.delete methods